feat: add extension methods for Android Uri ✨
To make the conversion from Uri to absolute path reusable
